This commit is wip. There are many things to fix here, but it shows what I want: 1. Make LiveData purely as state 2. Make Repository clean and stay away from keeping a reference to LiveData( we use MediatorLiveData cause when no on is obseving it, it'll be freed) 3. Move the heavily lifting query to ViewModel, or into a Presenter in the future. 4. The boradcast receiver should just update liveData, and as a source of Repository's Transformation.
